Fix runtime exceptions and update totals on "Editing Did End"
Runtime exceptions were raised when deleting all text from a field and
then switching to another field, because force unwrapping was used on
nil values. Optional binding was used to correct this issue.

After deleting all text from a field, the amounts were not updated with
the default values, leaving stale amounts. Added calls to tallyBill in
the IBAction's linked as "Editing Did End" callbacks.

Note: The call signature to tallyBill was changed; there was no need
to accept AnyObject, as all data is extracted from the text fields in
tallyBill.